refactor: validation

+++ services/reservationService.js
@@ -7,18 +7,21 @@ const createReservation = async (reservationData) => {
     !reservationData.checkout ||
     !reservationData.name
   ) {
-    throw new Error("Missing required fields: room, checkin, checkout, name");
+    throw new AppError(
+      "Missing required fields: room, checkin, checkout, name",
+      400
+    );
   }
 
   const checkinDate = new Date(reservationData.checkin);
   const checkoutDate = new Date(reservationData.checkout);
 
   if (isNaN(checkinDate.getTime()) || isNaN(checkoutDate.getTime())) {
-    throw new Error("Invalid check-in or check-out date format.");
+    throw new AppError("Invalid check-in or check-out date format.", 400);
   }
 
   if (checkinDate >= checkoutDate) {
-    throw new Error("Check-out date must be after check-in date.");
+    throw new AppError("Check-out date must be after check-in date.", 400);
   }
 
   return await reservationRepository.createReservation(reservationData);
@@ -28,7 +31,7 @@ const getAllReservations = async (query) => {
   const reservations = await reservationRepository.getAllReservations(query);
 
   if (reservations.length === 0) {
-    throw new Error("No reservations found matching the criteria.");
+    throw new AppError("No reservations found matching the criteria.", 404);
   }
 
   return reservations;
@@ -40,13 +43,21 @@ const getReservationById = async (reservationId) => {
   );
 
   if (!reservation) {
-    throw new Error("No reservation found with this ID.");
+    throw new AppError("No reservation found with this ID.", 404);
   }
 
   return reservation;
 };
 
 const updateReservation = async (reservationId, updateData) => {
+  const reservation = await reservationRepository.getReservationById(
+    reservationId
+  );
+
+  if (!reservation) {
+    throw new AppError("No reservation found with this ID.", 404);
+  }
+
   return await reservationRepository.updateReservation(
     reservationId,
     updateData
@@ -54,6 +65,14 @@ const updateReservation = async (reservationId, updateData) => {
 };
 
 const deleteReservation = async (reservationId) => {
+  const reservation = await reservationRepository.getReservationById(
+    reservationId
+  );
+
+  if (!reservation) {
+    throw new AppError("No reservation found with this ID.", 404);
+  }
+
   return await reservationRepository.deleteReservation(reservationId);
 };
